Add comment about casting to challenge

+++ ByteShortInt/src/com/dankb/Main.java
@@ -7,7 +7,9 @@ public class Main {
         byte byteChallengeValue = 120;
         short shortChallengeValue = 12000;
         int intChallengeValue = 1_200_450;
-        long longChallengeValue = (long) 50000 + (10* (byteChallengeValue + shortChallengeValue + intChallengeValue));
+
+        //not necessary to cast because long will accept an integer
+        long longChallengeValue =  50000 + (10* (byteChallengeValue + shortChallengeValue + intChallengeValue));
         System.out.println("The total is: " + longChallengeValue);
 
     }
